feat: Refactor achievement-tracker.test.ts into smaller files following single-responsibility principle
Split the 885-line achievement-tracker.test.ts into five single-responsibility test files that mirror the already-split source modules (performance, unlock/retirement, status, ladder), with shared fixtures extracted to a helper file. Pure test-only refactor with no source or behavior changes.

Changes:
- src/lib/achievement-fixtures.ts — new shared test fixtures (mockExercises array + createHistoryEntry helper), not a .test file so Vitest won't collect it
- src/lib/exercise-performance.test.ts — new: getBestPerformance tests plus the two performance edge cases (6 tests)
- src/lib/exercise-unlock-tracker.test.ts — new: isExerciseUnlocked, shouldRetireExercise, checkWorkoutAchievements, Dead Bug retirement, combined unlock/retirement (26 tests)
- src/lib/exercise-status.test.ts — new: getAvailableExercises, unlock chains, getExerciseStatuses, and three status edge cases (14 tests)
- src/lib/ladder-advancement.test.ts — new: checkLadderAdvancements with its inline local helpers (6 tests)
- src/lib/achievement-tracker.test.ts — rewritten from 885 lines to a lean barrel smoke test asserting every public symbol is re-exported (1 test)
- .dead-code-exceptions.json — added src/lib/achievement-fixtures.ts to files (imported only by excluded test files)
